nfc-backend/requirements.txt
pn532==1.5
